Code refactor.
1. Rename --degenerate into --anti_degenerate.
2. Move our code from loader.py into code_convertion.py.
