+++ src/data/loader.py
@@ -6,6 +6,7 @@
 #
 
 from logging import getLogger
+from .code_convertion import load_dict
 import os
 import numpy as np
 import torch
@@ -227,54 +228,6 @@ def load_para_data(params, data):
     logger.info("")
 
 
-def load_para_dict(filename):
-    para_dict = {}
-    dict_file = open(filename, mode="r", encoding="UTF-8")
-    lines = dict_file.readlines()
-    for line in lines:
-        line = line.strip()
-        if len(line) != 0:
-            tmp = line.split()
-            if tmp[0] in para_dict.keys():
-                para_dict[tmp[0]].append(tmp[1])
-            else:
-                para_dict[tmp[0]] = [tmp[1]]
-    dict_file.close()
-    logger.info("Read %d words from the dictionary file." % len(lines))
-
-    return para_dict
-
-
-def load_dict(params, data):
-    assert len(params.langs) == 2, "Need src and tgt two languages"
-
-    src_dict_path = os.path.join(params.data_path, "vocab.%s" % params.langs[0])
-    tgt_dict_path = os.path.join(params.data_path, "vocab.%s" % params.langs[1])
-    src_para_dict_path = os.path.join(params.data_path,
-                                      "dict.%s-%s.%s" % (params.langs[0], params.langs[1], params.langs[0]))
-    tgt_para_dict_path = os.path.join(params.data_path,
-                                      "dict.%s-%s.%s" % (params.langs[0], params.langs[1], params.langs[1]))
-
-    assert os.path.isfile(src_dict_path) and os.path.isfile(tgt_dict_path) \
-           and os.path.isfile(src_para_dict_path) and os.path.isfile(tgt_para_dict_path)
-
-    from .dictionary import Dictionary
-
-    logger.info("Read source monolingual vocabulary...")
-    src_vocab = Dictionary.read_vocab(src_dict_path)
-    data["src_vocab"] = src_vocab
-
-    logger.info("Read target monolingual vocabulary...")
-    tgt_vocab = Dictionary.read_vocab(tgt_dict_path)
-    data["tgt_vocab"] = tgt_vocab
-
-    logger.info("Read source parallel dictionary...")
-    data["src_para_dict"] = load_para_dict(src_para_dict_path)
-
-    logger.info("Read target parallel dictionary...")
-    data["tgt_para_dict"] = load_para_dict(tgt_para_dict_path)
-
-
 def check_data_params(params):
     """
     Check datasets parameters.
@@ -377,7 +330,7 @@ def load_data(params):
     load_para_data(params, data)
 
     # monolingual dict
-    if params.degenerate:
+    if params.anti_degenerate:
         load_dict(params, data)
 
     # monolingual data summary
